--- client/src/04 - utils/PrivateRoute.js
@@ -1,1 +1,0 @@
-import { Route, Redirect } from "react-router-dom";
